automation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pyautogui==0.9.54
openpyxl==3.1.2
pandas==2.1.3
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
typing-extensions==4.8.0